+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # CHANGELOG
+## 0.0.3
+### Add support for alternate palettes
+- Support for designating a PNG as the alternate palette
+- Support for dynamicaly changing to different palette
 ## 0.0.2
-### Add complete support for MagicaVoxel model format
+### Add support for loading all model in vox file
 - Support loading all models in the file
 - Cache mesh and voxel data in asset 
 ## 0.0.1